increase test coverage for MvCommand to 91% and RmCommand to 88%

+++ src/interpreter/commands/RmCommand.java
@@ -40,6 +40,9 @@ public class RmCommand {
     if (file.isDirectory()) {
       return "Error: Could not remove directory: " + file.getPath();
     } else if (file.isFile()) {
+      if (!file.canWrite()) {
+        return "Error: Could not remove file: " + file.getPath();
+      }
       return file.delete() ? "Removed file: " + file.getPath() : "Error: Could not remove file: " + file.getPath();
     }
 
@@ -53,5 +56,4 @@ public class RmCommand {
     }
     return file;
   }
-
 }
